EditDialog: refactor useHandleItemClick()
